Silence a build warning on %d for g_pgrp, cast to int

+++ stress-ng.c
@@ -1919,7 +1919,7 @@ static void kill_procs(const int sig)
 	/* multiple calls will always fallback to SIGKILL */
 	count++;
 	if (count > 5) {
-		pr_dbg("killing process group %d with SIGKILL\n", g_pgrp);
+		pr_dbg("killing process group %d with SIGKILL\n", (int)g_pgrp);
 		signum = SIGKILL;
 	}
